Add source based routing config gen test for more coverage (#270)
* More coverage for config gen

* rename sds-version

* restructure kube tests

* Use different header values
