+++ platform/kube/BUILD
@@ -45,9 +45,9 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
+        "client_test.go",
         "controller_test.go",
         "conversion_test.go",
-        "kube_test.go",
         "queue_test.go",
     ],
     data = [":kubeconfig"],
