+++ proxy/envoy/discovery_test.go
@@ -81,7 +81,7 @@ func TestServiceDiscoveryVersion(t *testing.T) {
 	url := "/v1/registration/" + mock.HelloService.Key(mock.HelloService.Ports[0],
 		map[string]string{"version": "v1"})
 	response := makeDiscoveryRequest(ds, url, t)
-	compareResponse(response, "testdata/sds-version.json", t)
+	compareResponse(response, "testdata/sds-v1.json", t)
 }
 
 func TestServiceDiscoveryEmpty(t *testing.T) {
@@ -94,7 +94,7 @@ func TestServiceDiscoveryEmpty(t *testing.T) {
 func TestClusterDiscovery(t *testing.T) {
 	registry := mock.MakeRegistry()
 	ds := makeDiscoveryService(registry)
-	url := fmt.Sprintf("/v1/clusters/%s/%s", IstioServiceCluster, mock.HostInstance)
+	url := fmt.Sprintf("/v1/clusters/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
 	response := makeDiscoveryRequest(ds, url, t)
 	compareResponse(response, "testdata/cds.json", t)
 }
@@ -103,23 +103,26 @@ func TestClusterDiscoveryCircuitBreaker(t *testing.T) {
 	registry := mock.MakeRegistry()
 	addCircuitBreaker(registry, t)
 	ds := makeDiscoveryService(registry)
-	url := fmt.Sprintf("/v1/clusters/%s/%s", IstioServiceCluster, mock.HostInstance)
+	url := fmt.Sprintf("/v1/clusters/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
 	response := makeDiscoveryRequest(ds, url, t)
 	compareResponse(response, "testdata/cds-circuit-breaker.json", t)
 }
 
 func TestRouteDiscovery(t *testing.T) {
 	ds := makeDiscoveryService(mock.MakeRegistry())
-	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstance)
+	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
 	response := makeDiscoveryRequest(ds, url, t)
-	compareResponse(response, "testdata/rds.json", t)
+	compareResponse(response, "testdata/rds-v0.json", t)
+	url = fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV1)
+	response = makeDiscoveryRequest(ds, url, t)
+	compareResponse(response, "testdata/rds-v1.json", t)
 }
 
 func TestRouteDiscoveryTimeout(t *testing.T) {
 	registry := mock.MakeRegistry()
 	addTimeout(registry, t)
 	ds := makeDiscoveryService(registry)
-	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstance)
+	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
 	response := makeDiscoveryRequest(ds, url, t)
 	compareResponse(response, "testdata/rds-timeout.json", t)
 }
@@ -128,7 +131,22 @@ func TestRouteDiscoveryWeighted(t *testing.T) {
 	registry := mock.MakeRegistry()
 	addWeightedRoute(registry, t)
 	ds := makeDiscoveryService(registry)
-	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstance)
+	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
 	response := makeDiscoveryRequest(ds, url, t)
 	compareResponse(response, "testdata/rds-weighted.json", t)
+}
+
+func TestRouteDiscoveryFault(t *testing.T) {
+	registry := mock.MakeRegistry()
+	addFaultRoute(registry, t)
+	ds := makeDiscoveryService(registry)
+
+	// fault rule is source based: we check that the rule only affect v0 and not v1
+	url := fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV0)
+	response := makeDiscoveryRequest(ds, url, t)
+	compareResponse(response, "testdata/rds-fault.json", t)
+
+	url = fmt.Sprintf("/v1/routes/80/%s/%s", IstioServiceCluster, mock.HostInstanceV1)
+	response = makeDiscoveryRequest(ds, url, t)
+	compareResponse(response, "testdata/rds-v1.json", t)
 }
