+++ proxy/envoy/testdata/envoy-v1.json.golden
@@ -1,0 +1,264 @@
+{
+  "listeners": [
+    {
+      "port": 80,
+      "filters": [
+        {
+          "type": "read",
+          "name": "http_connection_manager",
+          "config": {
+            "codec_type": "auto",
+            "stat_prefix": "http",
+            "rds": {
+              "cluster": "rds",
+              "route_config_name": "80",
+              "refresh_delay_ms": 1000
+            },
+            "filters": [
+              {
+                "type": "both",
+                "name": "mixer",
+                "config": {
+                  "mixer_server": "mixer:9091"
+                }
+              },
+              {
+                "type": "decoder",
+                "name": "router",
+                "config": {}
+              }
+            ],
+            "access_log": [
+              {
+                "path": "/dev/stdout"
+              }
+            ]
+          }
+        }
+      ],
+      "bind_to_port": false
+    },
+    {
+      "port": 81,
+      "filters": [
+        {
+          "type": "read",
+          "name": "http_connection_manager",
+          "config": {
+            "codec_type": "auto",
+            "stat_prefix": "http",
+            "rds": {
+              "cluster": "rds",
+              "route_config_name": "81",
+              "refresh_delay_ms": 1000
+            },
+            "filters": [
+              {
+                "type": "both",
+                "name": "mixer",
+                "config": {
+                  "mixer_server": "mixer:9091"
+                }
+              },
+              {
+                "type": "decoder",
+                "name": "router",
+                "config": {}
+              }
+            ],
+            "access_log": [
+              {
+                "path": "/dev/stdout"
+              }
+            ]
+          }
+        }
+      ],
+      "bind_to_port": false
+    },
+    {
+      "port": 90,
+      "filters": [
+        {
+          "type": "read",
+          "name": "tcp_proxy",
+          "config": {
+            "stat_prefix": "tcp",
+            "route_config": {
+              "routes": [
+                {
+                  "cluster": "inbound:1090",
+                  "destination_ip_list": [
+                    "10.1.0.0/32"
+                  ]
+                },
+                {
+                  "cluster": "outbound:world.default.svc.cluster.local:custom",
+                  "destination_ip_list": [
+                    "10.2.0.0/32"
+                  ],
+                  "destination_ports": "90"
+                }
+              ]
+            }
+          }
+        }
+      ],
+      "bind_to_port": false
+    },
+    {
+      "port": 1081,
+      "filters": [
+        {
+          "type": "read",
+          "name": "http_connection_manager",
+          "config": {
+            "codec_type": "auto",
+            "stat_prefix": "http",
+            "rds": {
+              "cluster": "rds",
+              "route_config_name": "1081",
+              "refresh_delay_ms": 1000
+            },
+            "filters": [
+              {
+                "type": "both",
+                "name": "mixer",
+                "config": {
+                  "mixer_server": "mixer:9091"
+                }
+              },
+              {
+                "type": "decoder",
+                "name": "router",
+                "config": {}
+              }
+            ],
+            "access_log": [
+              {
+                "path": "/dev/stdout"
+              }
+            ]
+          }
+        }
+      ],
+      "bind_to_port": false
+    },
+    {
+      "port": 1090,
+      "filters": [
+        {
+          "type": "read",
+          "name": "tcp_proxy",
+          "config": {
+            "stat_prefix": "tcp",
+            "route_config": {
+              "routes": [
+                {
+                  "cluster": "inbound:1090",
+                  "destination_ip_list": [
+                    "10.1.1.1/32"
+                  ]
+                }
+              ]
+            }
+          }
+        }
+      ],
+      "bind_to_port": false
+    },
+    {
+      "port": 5001,
+      "filters": [],
+      "bind_to_port": true,
+      "use_original_dst": true
+    }
+  ],
+  "admin": {
+    "access_log_path": "/dev/stdout",
+    "port": 5000
+  },
+  "cluster_manager": {
+    "clusters": [
+      {
+        "name": "inbound:1081",
+        "connect_timeout_ms": 1000,
+        "type": "static",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://127.0.0.1:1081"
+          }
+        ]
+      },
+      {
+        "name": "inbound:1090",
+        "connect_timeout_ms": 1000,
+        "type": "static",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://127.0.0.1:1090"
+          }
+        ]
+      },
+      {
+        "name": "inbound:80",
+        "connect_timeout_ms": 1000,
+        "type": "static",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://127.0.0.1:80"
+          }
+        ]
+      },
+      {
+        "name": "outbound:world.default.svc.cluster.local:custom",
+        "service_name": "world.default.svc.cluster.local:custom",
+        "connect_timeout_ms": 1000,
+        "type": "sds",
+        "lb_type": "round_robin"
+      },
+      {
+        "name": "rds",
+        "connect_timeout_ms": 1000,
+        "type": "strict_dns",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://manager:8080"
+          }
+        ]
+      }
+    ],
+    "sds": {
+      "cluster": {
+        "name": "sds",
+        "connect_timeout_ms": 1000,
+        "type": "strict_dns",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://manager:8080"
+          }
+        ]
+      },
+      "refresh_delay_ms": 1000
+    },
+    "cds": {
+      "cluster": {
+        "name": "cds",
+        "connect_timeout_ms": 1000,
+        "type": "strict_dns",
+        "lb_type": "round_robin",
+        "hosts": [
+          {
+            "url": "tcp://manager:8080"
+          }
+        ]
+      },
+      "refresh_delay_ms": 1000
+    }
+  }
+}
