+++ proxy/envoy/testdata/sds-v1.json.golden
